LayoutTests/fast/scrolling/mac/scrollend-event-user-scroll-basic-expected.txt


Test scroll over content
PASS overflowScrollendEventCount == 1 is true
PASS windowScrollendEventCount == 0 is true

Test scroll over document
PASS overflowScrollendEventCount == 0 is true
PASS windowScrollendEventCount == 1 is true
PASS successfullyParsed is true

TEST COMPLETE
